% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U.R
\name{diffintegrate}
\alias{diffintegrate}
\title{different way of integration
This function generates the values of integration.}
\usage{
diffintegrate(f, L, U, type = "simpson", n = 500, t = 50)
}
\arguments{
\item{f}{a function which you want to integrate.}

\item{L}{lower bound.}

\item{U}{upper bound.}

\item{type}{different way of integration "simpson"(default),"trapezoid","riemann"}

\item{n}{cutting into n pieces.}

\item{t}{simulating t times.}
}
\value{
a scalar equivalent to integrate from lower bound to upper bound.
}
\description{
different way of integration
This function generates the values of integration.
}
\examples{
## integrate from 1 to 2 by simpson method
f=function(x) 4*x^3
diffintegrate(f,1,2)

## integrate by trapezoid method
f=function(x) 4*x^3
diffintegrate(f,1,2,type="trapezoid",n=50,t=100)
}
\references{
http://www.tf.uns.ac.rs/~omorr/radovan_omorjan_003_prII/r-examples/spuRs/spuRs-2ed.pdf
}
\seealso{
\code{integrate}
}
